refactor: 경로를 /faqs로 변경하면서 컴포넌트 디렉토리 명도 함께 변경 (#117)
