+++ src/app/(root)/faqs/error.tsx
