+++ src/app/(root)/faqs/page.tsx
