+++ src/components/faqs/SearchedFaq.tsx
