+++ src/components/faqs/index.ts
