+++ src/components/index.ts
@@ -4,4 +4,4 @@ export * from './icons';
 export * from './products';
 export * from './auth';
 export * from './cart';
-export * from './faq';
+export * from './faqs';
